pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "capture-pdf"
version = "0.1.0"
description = "Capture pages from Kyobo Library app and compile to PDF"
requires-python = ">=3.10"
dependencies = [
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "Pillow>=10.0",
    "numpy>=1.24",
    "click>=8.0",
    "PyPDF2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "ruff>=0.1",
]

[project.scripts]
capture-pdf = "capture_pdf.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: marks tests as integration tests"]
